Delta versioon 3.13.16.18
Selle Delta versiooniga tulevad kaasa tugi 2.5/2.6 -> 3.13 üleminekuks ning parandused.

Antud Delta 3.13 versioon tuleb kas paigaldada nullist (s.t. tühja andmebaasi ja andmekausta peale) või uusima 2.5 versiooni peale või uusima 2.6 versiooni peale või eelmise 3.13 versiooni peale. Teiste versioonide peale paigaldamine ei ole toetatud.

Süsteemiadministraatorile: infra poolt puudutavad punktid, mida 3.6 versioonis ei ole ja tulevad alates 3.13 versioonist (sama nimekiri sisaldus ka eelmises 3.13 tarne kirjas):
* dir.root.shared ja dir.root.local konf.seaded ehk andmekausta jagamine kaheks osaks;
* monitooringu info teenus ehk /dhs/service/monitor;
* ldap.* ja kerberos.* konf.seaded ehk Active Directory'ga liidestus;
* logo.file konfiguratsiooniseade ehk custom logo kasutamine;
* OpenOffice'iga faili muutmiseks avamise funktsionaalsus (WebDAV protokolliga). (Teadaolev puudujääk: OpenOffice’iga faili muutmiseks avamisel ei tööta faili lukustamine.)
Paigaldusjuhendis on nimetatud punktide juurde lisatud ka märkus, et need 3.6 versioonis puuduvad.
Lisaks on suurendatud Delta sessiooni vaikimisi pikkust 3h peale.


Korrastada SIM ja SMIT alfresco-global.properties konf.failid järgnevalt:
* SIM alfresco-global.properties konf.fail:
  + lisada system.usages.enabled=false
  + vajadusel lisada useClientIpFromXForwardedForHttpHeader=true
  + eemaldada importer* seaded, kuna neid ei lähe vaja
  + eemaldada passthru.authentication.domain, kuna seda ei lähe vaja kui passthru.authentication.servers kasutatakse
* SMIT alfresco-global.properties konf.faili:
  + lisada system.usages.enabled=false
  + vajadusel lisada useClientIpFromXForwardedForHttpHeader=true
  + eemaldada importer* seaded, kuna neid ei lähe vaja
  + eemaldada passthru.authentication.domain, kuna seda ei lähe vaja kui passthru.authentication.servers kasutatakse
  + eemaldada lucene.query.maxClauses seade
  + eemaldada lõpust, pärast lucene.query.maxClauses, kõik seaded (*update*, import*, ...)


Üleminekuks uusimalt Delta 2.5 või uusimalt 2.6 versioonile 3.13 versioonile:
1) Eelnevalt seadistada OpenOffice käivituma soffice-monitor.sh skripti kaudu.
2) Ülemineku ajaks lisada alfresco-global.properties konf.faili read:
     updater.version25To313.enabled=true
     # xml fail, millele viidata, tuleb rakendusele loetavasse asukohta kopeerida https://svn.smit/tarkvara/delta/branches/develop-3.13/doc/delta2-migratsioon/ kaustast vastavalt asutusele
     updater.version25To313.documentTypes=/path/to/documentTypes_XXX.xml
     index.tracking.disableInTransactionIndexing=false
     findHolesAndIndex.enabled=false
     jobs.enabled=false
3) Veenduda, et ülemineku ajal oleks AMR teenus kättesaadav, kuna Delta tahab pärida struktuuriüksusi.
4) Käivitada Delta 3.13 versioon; käivitamisel automaatselt:
     1. Luuakse ${dir.root} kausta alamkaustad shared ja local
     2. Liigutatakse ${dir.root} sees olevad kaustad ${dir.root.shared} ja ${dir.root.local} alamkaustadesse (vastavalt  2. alampunktis toodud jaotusele); luuakse ${dir.root.shared}/updater kaust ja liigutatakse sellesse kõik ${dir.root}/*.csv failid. Arvestatakse, et ühegi alamkausta asukohta ei ole konfis eraldi kästisi määratud (s.t ainsana on määratud dir.root). Kui see nii ei ole, siis tuleb enne 3.13 versiooni esmakordset käivitamist:
       * Vajadusel määrata konfis dir.root.shared ja dir.root.local väärtused; määramata jätmisel on vaikeväärtused dir.root.shared=${dir.root}/shared ja dir.root.local=${dir.root}/local
       * Veenduda, et järgnevad kaustad oleksid rakendusele kättesaadavad järgnevates asukohtades (vajadusel liigutada nendesse asukohtadesse):
         + ${dir.root.local}/lucene-indexes
         + ${dir.root.local}/backup-lucene-indexes  (ei ole kohustuslik et eksisteeriks)
         + ${dir.root.local}/oouser  (ei ole kohustuslik et eksisteeriks)
         + ${dir.root.shared}/audit.contentstore
         + ${dir.root.shared}/contentstore
         + ${dir.root.shared}/contentstore.deleted
         + ${dir.root.shared}/updater  (ei ole kohustuslik et eksisteeriks)
5) Käivitumisel sooritatakse muudatused dokumentide andmetes ja muudes andmetes. Kui üleminek lõpetab, siis rakendus lõpetab deploymise ja muutub veebist kättesaadavaks.
6) Pärast ülemineku edukat lõpetamist kontrollida alfresco.log failist, kas ülemineku ajal oli logis mõni exception. Üle vaadata ${dir.root.shared}/updater kaustast convertToDynamicDocumentsUpdaterFailed.csv (kõik read) ja documentInvalidAccessRestrictionUpdaterCompleted.csv (read mis sisaldavad notValidAndUnchanged).
7) Pärast ülemineku edukat lõpetamist eemaldada alfresco-global.properties konf.failist punktis 2 lisatud read.
8) Pärast ülemineku edukat lõpetamist - toodangkeskkonnas on soovitav teha seda väljaspool tööaega - minna Administraator -> Nodeide sirvija ning:
     1. Vajutada "Käivita tööülesannete kustutamise skript". See on lõpetanud siis, kui alfresco.log faili tuleb rida, mis sisaldab fraasi "DeleteAllTasksFromRepo] Completed nodes updater".
     2. Kui eelnev on lõpetanud, siis vajutada "Käivita logikirjete kustutamise skript". See on lõpetanud siis, kui alfresco.log faili tuleb rida, mis sisaldab fraasi "DeleteAllLogNodesFromRepo] Completed nodes updater".

Testisime üleminekut keskkonnas, kus:
* Rakendusserver: 4 Xeon CPU-d ja 32 GB RAM ning lucene-indexes kaust oli RAM kettal (nimetatud RAM hulga sees)
     + Java võtmed olid SIM puhul -XX:MaxPermSize=192m -Xmx8192m -XX:NewRatio=2 -server -d64
     + Java võtmed olid SMIT puhul -XX:MaxPermSize=192m -Xmx4096m -XX:NewRatio=2 -server -d64
* Andmebaasiserver: 4 Xeon CPU-d ja 32 GB RAM
Üleminek võttis aega:
* SIM - 40 tundi
* SMIT - 5,5 tundi


git-svn-id: https://svn.smit/tarkvara/delta/branches/develop-3.13@129337 5fddf92a-9921-11dd-8d32-b999f448c456

+++ source/java/ee/webmedia/alfresco/docconfig/generator/systematic/AccessRestrictionGenerator.java
@@ -34,13 +34,14 @@ import org.apache.commons.lang.time.DateUtils;
 import ee.webmedia.alfresco.adr.service.AdrService;
 import ee.webmedia.alfresco.classificator.enums.AccessRestriction;
 import ee.webmedia.alfresco.classificator.enums.PublishToAdr;
-import ee.webmedia.alfresco.common.model.DynamicBase;
 import ee.webmedia.alfresco.classificator.model.ClassificatorValue;
 import ee.webmedia.alfresco.classificator.service.ClassificatorService;
+import ee.webmedia.alfresco.common.model.DynamicBase;
 import ee.webmedia.alfresco.common.propertysheet.classificatorselector.ClassificatorSelectorAndTextGenerator;
 import ee.webmedia.alfresco.common.propertysheet.config.WMPropertySheetConfigElement.ItemConfigVO;
 import ee.webmedia.alfresco.common.propertysheet.multivalueeditor.PropsBuilder;
 import ee.webmedia.alfresco.common.web.BeanHelper;
+import ee.webmedia.alfresco.common.web.WmNode;
 import ee.webmedia.alfresco.docadmin.service.DocumentAdminService;
 import ee.webmedia.alfresco.docadmin.service.Field;
 import ee.webmedia.alfresco.docadmin.service.FieldGroup;
@@ -63,6 +64,8 @@ import ee.webmedia.alfresco.utils.TextUtil;
  * @author Alar Kvell
  */
 public class AccessRestrictionGenerator extends BaseSystematicFieldGenerator {
+
+    public static final String BEAN_NAME = "accessRestrictionGenerator";
 
     private static final String VIEW_MODE_PROP_SUFFIX = "View";
     public static final String ACCESS_RESTRICTION_CHANGE_REASON_ERROR = "accessRestrictionChangeReasonError";
@@ -376,10 +379,12 @@ public class AccessRestrictionGenerator extends BaseSystematicFieldGenerator {
         // If user changed the access restriction, verify that reason was also changed
         final String reason = (String) document.getProp(DocumentDynamicDialog.TEMP_ACCESS_RESTRICTION_CHANGE_REASON);
         if (StringUtils.isNotBlank(reason)) {
-            // Reset the reason in repository so DocumentPropertyChangeHolder can pick up the change when user enters the same value
-            nodeService.removeProperty(nodeRef, ACCESS_RESTRICTION_CHANGE_REASON);
-            document.setProp(ACCESS_RESTRICTION_CHANGE_REASON, reason);
-            document.getNode().getProperties().remove(DocumentDynamicDialog.TEMP_ACCESS_RESTRICTION_CHANGE_REASON);
+            if (!Boolean.TRUE.equals(document.getProp(DocumentDynamicDialog.TEMP_VALIDATE_WITHOUT_SAVE))) {
+                // Reset the reason in repository so DocumentPropertyChangeHolder can pick up the change when user enters the same value
+                nodeService.removeProperty(nodeRef, ACCESS_RESTRICTION_CHANGE_REASON);
+                document.setProp(ACCESS_RESTRICTION_CHANGE_REASON, reason);
+                document.getNode().getProperties().remove(DocumentDynamicDialog.TEMP_ACCESS_RESTRICTION_CHANGE_REASON);
+            }
             return;
         }
 
@@ -421,17 +426,29 @@ public class AccessRestrictionGenerator extends BaseSystematicFieldGenerator {
 
         String oldAccessRestriction = (String) oldProps.get(ACCESS_RESTRICTION);
         if (oldAccessRestriction != null && !accessRestriction.equals(oldAccessRestriction)
-                    && (AccessRestriction.INTERNAL.equals(accessRestriction) || AccessRestriction.OPEN.equals(accessRestriction))) {
-            newProps.put(ACCESS_RESTRICTION_REASON.toString(), null);
-            newProps.put(ACCESS_RESTRICTION_BEGIN_DATE.toString(), null);
-            newProps.put(ACCESS_RESTRICTION_END_DATE.toString(), null);
-            newProps.put(ACCESS_RESTRICTION_END_DESC.toString(), null);
+                && (AccessRestriction.INTERNAL.equals(accessRestriction) || AccessRestriction.OPEN.equals(accessRestriction))) {
+            setHiddenFieldsNull(newProps);
         }
 
         if (!document.isDraftOrImapOrDvk()) {
             if (!getChangedAccessRestrictionFieldIds(document, oldProps).isEmpty()) {
                 document.setAccessRestrictionPropsChanged(true);
             }
+        }
+    }
+
+    private void setHiddenFieldsNull(Map<String, Object> newProps) {
+        newProps.put(ACCESS_RESTRICTION_REASON.toString(), null);
+        newProps.put(ACCESS_RESTRICTION_BEGIN_DATE.toString(), null);
+        newProps.put(ACCESS_RESTRICTION_END_DATE.toString(), null);
+        newProps.put(ACCESS_RESTRICTION_END_DESC.toString(), null);
+    }
+
+    public void clearHiddenValues(WmNode node) {
+        Map<String, Object> properties = node.getProperties();
+        String accessRestriction = (String) properties.get(ACCESS_RESTRICTION);
+        if (AccessRestriction.INTERNAL.equals(accessRestriction) || AccessRestriction.OPEN.equals(accessRestriction)) {
+            setHiddenFieldsNull(properties);
         }
     }
